requirements.txt
# Core dependencies
anthropic>=0.39.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Web scraping (for job postings)
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0

# Data handling
pandas>=2.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0
